% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aux_maps.R
\name{get_aux_maps}
\alias{get_aux_maps}
\title{Get auxiliary phenotype ontology maps}
\usage{
get_aux_maps(
  cache_dir = NA,
  force_download = FALSE,
  primary_site = NA,
  ignore_minor_type = FALSE
)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should be overwritten
(set to TRUE to re-download if file exists in cache)}
}
\value{
\strong{metadata} - A data frame with 4 rows and 6 columns:
\itemize{
\item \emph{source} - gene annotation source
\item \emph{annotation_data} - type of annotations used
\item \emph{url} - URL of annotation resource
\item \emph{citation} - publication to cite for annotation source
(citation; PMID)
\item \emph{version} - version used
\item \emph{abbreviation} - abbreviation used in column names of records
}

\strong{records} - A list with the following elements:
\itemize{
\item \emph{umls} - List with UMLS data dictionaries
\item \emph{do} - Data frame with DO identifiers/names
\item \emph{efo} - List with EFO data dictionaries
\item \emph{icd10} - Data frame with CUI to ICD10 cross-reference
}
}
\description{
Downloads and returns multiple phenotype dictionaries used for
cross-referencing of terms, including UMLS, EFO, DO, and ICD10.
The dataset comes as a \code{list} object, with two elements:
\itemize{
\item \code{metadata} - a data frame with metadata regarding annotation
resources used
\item \code{records} - a list object with multiple lists/data.frames
}
}
\examples{

\dontrun{
library(geneOncoX)
oncology_terms <- get_aux_maps(cache_dir = tempdir())
}

}
