% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_oncox.R
\name{get_pox_data}
\alias{get_pox_data}
\title{Function that retrieves phenOncoX data from Google Drive/local cache}
\usage{
get_pox_data(cache_dir = NA, force_download = FALSE, db = "oncotree_core")
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache
should be overwritten (set to TRUE to re-download if file
exists in cache)}

\item{db}{type of dataset to be retrieved}
}
\value{
pre-processed ontology terms records/metadata or auxiliary maps
}
\description{
Function that retrieves phenOncoX data from Google Drive/local cache
}
\keyword{internal}
