% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Get core phenotype terms from OncoTree}
\usage{
get_tree(
  cache_dir = NA,
  force_download = FALSE,
  primary_site = NA,
  max_tree_depth = 6,
  ignore_minor_type = FALSE
)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should be overwritten
(set to TRUE to re-download if file exists in cache)}

\item{max_tree_depth}{consider only terms up to a given depth in
the OncoTree (integer from 1-6, default: 6)}

\item{ignore_minor_type}{logical indicating if minor tumor types
should be excluded or not}

\item{site}{limit OncoTree terms to those relevant for a given
primary tumor type/site. Possible values:
'Adrenal Gland', 'Ampulla of Vater', 'Biliary Tract', 'Bladder/Urinary Tract',
'Bone', 'Breast', 'Cervix', 'CNS/Brain', 'Colon/Rectum', 'Esophagus/Stomach',
'Eye', 'Head and Neck', 'Kidney', 'Liver', 'Lung', 'Lymphoid', 'Myeloid',
'Other/Unknown', 'Ovary/Fallopian Tube', 'Pancreas', 'Penis',
'Peripheral Nervous System', 'Peritoneum', 'Pleura', 'Prostate',
'Skin', 'Soft Tissue', 'Testis', 'Thymus', 'Thyroid', 'Uterus',
'Vulva/Vagina'}
}
\value{
\strong{metadata} - A data frame with 1 row and 6 columns:
\itemize{
\item \emph{source} - gene annotation source
\item \emph{annotation_data} - type of annotations used
\item \emph{url} - URL of annotation resource
\item \emph{citation} - publication to cite for annotation source
(citation; PMID)
\item \emph{version} - version used
\item \emph{abbreviation} - abbreviation used in column names of records
}

\strong{records} - A data frame with the following columns:
\itemize{
\item \emph{primary_site} - Primary tumor type/site
\item \emph{ot_main_type} - Main tumor type (OncoTree)
\item \emph{ot_name} - Phenotype name (OncoTree)
\item \emph{ot_level} - Tree level (OncoTree)
\item \emph{ot_code} - Phenotype code (OncoTree)
\item \emph{ot_code_path} - Tree code path (OncoTree)
\item \emph{cui} - Concept unique identifier (UMLS)
\item \emph{cui_name} - Phenotype name (UMLS)
\item \emph{minor_type} - logical indicating whether the term is part
of a minor tumor type/site
\item \emph{source} - term indicating the dataset source
}
}
\description{
Downloads and returns a dataset with curated phenotype terms (UMLS) from
OncoTree. The dataset comes as a \code{list} object, with two elements:
\itemize{
\item \code{metadata} - a data frame with metadata regarding annotation
resources used
\item \code{records} - a data frame with phenotype terms
}
}
\examples{

\dontrun{
library(geneOncoX)
oncology_terms <- get_tree(cache_dir = tempdir())
}

}
